JIRAPLUGIN-191

+++ src/main/java/org/everit/jira/timetracker/plugin/JiraTimetrackerPluginImpl.java
@@ -287,7 +287,7 @@ public class JiraTimetrackerPluginImpl implements JiraTimetrackerPlugin, Initial
    */
   private TimeTrackingConfiguration timeTrackingConfiguration;
 
-  private long workHoursPerDay;
+  private double workHoursPerDay;
 
   /**
    * Default constructor.
@@ -301,7 +301,7 @@ public class JiraTimetrackerPluginImpl implements JiraTimetrackerPlugin, Initial
   @Override
   public void afterPropertiesSet() throws Exception {
     generatePluginUUID();
-    workHoursPerDay = timeTrackingConfiguration.getHoursPerDay().longValue();
+    workHoursPerDay = timeTrackingConfiguration.getHoursPerDay().doubleValue();
 
     loadJttpBuildProperties();
 
@@ -824,7 +824,7 @@ public class JiraTimetrackerPluginImpl implements JiraTimetrackerPlugin, Initial
       timeSpent += worklog.getLong("timeworked").longValue();
     }
 
-    long expectedTimeSpent = workHoursPerDay * DateTimeConverterUtil.SECONDS_PER_MINUTE
+    double expectedTimeSpent = workHoursPerDay * DateTimeConverterUtil.SECONDS_PER_MINUTE
         * DateTimeConverterUtil.MINUTES_PER_HOUR;
     return timeSpent >= expectedTimeSpent;
   }
